# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AIPrintVerse API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signin:
    post:
      operationId: signIn
      tags: [auth]
      summary: Sign in with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignInInput"
      responses:
        "200":
          description: Session token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid credentials
  /auth/signup:
    post:
      operationId: signUp
      tags: [auth]
      summary: Sign up with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignUpInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
...
  /auth/roles:
    get:
      operationId: getUserRoles
      tags: [auth]
      summary: List user roles (admin only)
      responses:
        "200":
          description: List of user roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserRole"
    post:
      operationId: assignRole
      tags: [auth]
      summary: Assign role to user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleInput"
      responses:
        "201":
          description: Role assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRole"

  # Designs
  /designs:
    get:
      operationId: listDesigns
      tags: [designs]
      summary: List designs with optional category filter
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of designs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DesignListResponse"
    post:
      operationId: createDesign
      tags: [designs]
      summary: Create a design (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DesignInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Design"
  /designs/{id}:
    get:
      operationId: getDesign
      tags: [designs]
      summary: Get design by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Design
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Design"
        "404":
          description: Not found
    patch:
      operationId: updateDesign
      tags: [designs]
      summary: Update design (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DesignUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Design"
    delete:
      operationId: deleteDesign
      tags: [designs]
      summary: Delete design (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /blog/posts:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List published blog posts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: pageSize
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated blog posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPostListResponse"
    post:
      operationId: createBlogPost
      tags: [blog]
      summary: Create blog post (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /blog/posts/{id}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get blog post by ID or slug
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found
    patch:
      operationId: updateBlogPost
      tags: [blog]
      summary: Update blog post (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    delete:
      operationId: deleteBlogPost
      tags: [blog]
      summary: Delete blog post (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /blog/posts/slug/{slug}:
    get:
      operationId: getBlogPostBySlug
      tags: [blog]
      summary: Get published blog post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found
  /blog/categories:
    get:
      operationId: listBlogCategories
      tags: [blog]
      summary: List blog categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogCategory"
    post:
      operationId: createBlogCategory
      tags: [blog]
      summary: Create blog category (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogCategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogCategory"
...
components:
  schemas:
...
    UserInfo:
      type: object
      properties:
        id:
          type: string
        email:
          type: ["string", "null"]
        isAdmin:
          type: boolean
      required: [id, isAdmin]
    SignInInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    SignUpInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthSession:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserInfo"
      required: [token, user]
    UserRole:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        role:
          type: string
        created_at:
          type: string
      required: [id, user_id, role, created_at]
    UserRoleInput:
      type: object
      properties:
        user_id:
          type: string
        role:
          type: string
      required: [user_id, role]

    # Designs
    Design:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        image_url:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        teepublic_url:
          type: ["string", "null"]
        redbubble_url:
          type: ["string", "null"]
        amazon_url:
          type: ["string", "null"]
        etsy_url:
          type: ["string", "null"]
        featured:
          type: boolean
        source:
          type: ["string", "null"]
        external_id:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: string
      required: [id, name, image_url, category, featured]
    DesignInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        image_url:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        teepublic_url:
          type: string
        redbubble_url:
          type: string
        amazon_url:
          type: string
        etsy_url:
          type: string
        featured:
          type: boolean
        source:
          type: string
        external_id:
          type: string
      required: [name, image_url]
    DesignUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        image_url:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        teepublic_url:
          type: string
        redbubble_url:
          type: string
        amazon_url:
          type: string
        etsy_url:
          type: string
        featured:
          type: boolean
    DesignListResponse:
      type: object
      properties:
        designs:
          type: array
          items:
            $ref: "#/components/schemas/Design"
        total:
          type: integer
      required: [designs, total]
...
    BlogPost:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        featured_image:
          type: ["string", "null"]
        author_id:
          type: ["string", "null"]
        author_name:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        keywords:
          type: array
          items:
            type: string
        status:
          type: string
        read_time:
          type: ["string", "null"]
        meta_title:
          type: ["string", "null"]
        meta_description:
          type: ["string", "null"]
        related_designs:
          type: array
          items:
            type: string
        published_at:
          type: ["string", "null"]
        scheduled_publish_at:
          type: ["string", "null"]
        view_count:
          type: ["integer", "null"]
        source:
          type: string
        template_id:
          type: ["string", "null"]
        generation_batch:
          type: ["string", "null"]
        design_id:
          type: ["string", "null"]
        indexing_status:
          type: string
        ebook_ready:
          type: boolean
        created_at:
          type: string
        updated_at:
          type: string
      required: [id, title, slug, author_name, category, status]
    BlogPostInput:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        featured_image:
          type: string
        author_name:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        keywords:
          type: array
          items:
            type: string
        status:
          type: string
        read_time:
          type: string
        meta_title:
          type: string
        meta_description:
          type: string
        related_designs:
          type: array
          items:
            type: string
        published_at:
          type: string
        scheduled_publish_at:
          type: string
        source:
          type: string
        template_id:
          type: string
        generation_batch:
          type: string
        design_id:
          type: string
        indexing_status:
          type: string
      required: [title, slug]
    BlogPostUpdate:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        featured_image:
          type: string
        author_name:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        keywords:
          type: array
          items:
            type: string
        status:
          type: string
        read_time:
          type: string
        meta_title:
          type: string
        meta_description:
          type: string
        related_designs:
          type: array
          items:
            type: string
        published_at:
          type: string
        scheduled_publish_at:
          type: string
        indexing_status:
          type: string
        ebook_ready:
          type: boolean
    BlogPostListResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/BlogPost"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
      required: [posts, total, page, pageSize, totalPages]
    BlogCategory:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        created_at:
          type: string
      required: [id, name, slug, created_at]
    BlogCategoryInput:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
      required: [name, slug]
...