# lib/api-zod/src/generated/types/authSession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AIPrintVerse API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserInfo } from './userInfo';

export interface AuthSession {
  token: string;
  user: UserInfo;
}