# lib/api-zod/src/generated/types/userRole.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AIPrintVerse API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserRole {
  id: string;
  user_id: string;
  role: string;
  created_at: string;
}